main_access/requirements.txt
dbus-python==1.2.12
decorator==4.4.1
Django==2.2.6
django-console==0.4.7
django-rest-framework==0.1.0
djangorestframework==3.10.3
evento==1.0.1
gunicorn==20.0.4
omxplayer-wrapper==0.3.2
paho-mqtt==1.5.0
pathlib2==2.3.5
pkg-resources==0.0.0
psycopg2==2.8.4
pyserial==3.4
pytz==2019.3
redcmd==1.2.10
redlib==1.5.7
RPi.GPIO==0.7.0
six==1.12.0
sqlparse==0.3.0
vlc-ctrl==1.0.8
zope.interface==4.6.0
